Fix broken protobuf formula
I believe I'm currently having trouble due to this. A package I'm working on's current CI
uses the 'develop' tag, and while it used to work, it's now broken. I think that what's happened
is that:

Protobuf consists of two parts:
1. The protoc compiler. In the proto distribution, this is held in 'bin/'.
2. The 'well known types' - things like 'Timestamp'. These are in 'include/'.

When proto was installed by untarring (pre #14) there were two tar lines, one
for each of the two folders. After #14, only protoc is laid onto the disk.

This PR adds a 'unix package' type of formula. All it does is lay down the
directories into /usr/local, similar perhaps to a `make install` command.

It tries to be as simple as possible.